Breakout the default config.
