+++ CMakeLists.txt
@@ -79,7 +79,8 @@ set(SOURCE_FILES
         RasterHists.cxx
         RasterEvioTool.cxx
         RasterMonEventInfo.cxx
-        RasterLogBookEntry.cxx)
+        RasterLogBookEntry.cxx
+        RasterMonDefaultConfig.cxx)
 
 add_library(RasterLib SHARED ${SOURCE_FILES} RasterLib_dict.cxx)
