+++ RasterHists.h
@@ -127,14 +127,4 @@ ClassDef(RasterHists, 0)
 
 };
 
-void Default_Setup_Raster_Tab(RasterHists *r);
-void Default_Setup_Raw_Raster_Tab(RasterHists *r);
-void Default_Setup_Raw_Raster_Tab2(RasterHists *r);
-void Default_Setup_Raw_Raster_Tab3(RasterHists *r);
-void Default_Setup_1_Channel_Scope(RasterHists *r);
-void Default_Setup_2_Channel_Scope(RasterHists *r);
-void Default_Setup_3_Channel_Scope(RasterHists *r);
-void Default_Setup_Helicity(RasterHists *r);
-void Default_Initialize_Histograms(RasterHists *r);
-
 #endif //RASTERMON_RASTERHISTS_H
