+++ RasterLibLinkDef.h
@@ -13,4 +13,13 @@
 #pragma link C++ struct Graph_t;
 #pragma link C++ struct Histogram_t;
 #pragma link C++ struct TabSpace_t;
+#pragma link C++ function Default_Setup_Raster_Tab;
+#pragma link C++ function Default_Setup_Raw_Raster_Tab;
+#pragma link C++ function Default_Setup_Raw_Raster_Tab2;
+#pragma link C++ function Default_Setup_Raw_Raster_Tab3;
+#pragma link C++ function Default_Setup_1_Channel_Scope;
+#pragma link C++ function Default_Setup_2_Channel_Scope;
+#pragma link C++ function Default_Setup_3_Channel_Scope;
+#pragma link C++ function Default_Setup_Helicity;
+#pragma link C++ function Default_Initialize_Histograms;
