+++ RasterMonDefaultConfig.h
@@ -1,0 +1,20 @@
+//
+// Created by Maurik Holtrop on 5/28/22.
+//
+
+#ifndef RASTERMON_RASTERMONDEFAULTCONFIG_H
+#define RASTERMON_RASTERMONDEFAULTCONFIG_H
+
+#include "RasterHists.h"
+
+void Default_Setup_Raster_Tab(RasterHists *r);
+void Default_Setup_Raw_Raster_Tab(RasterHists *r);
+void Default_Setup_Raw_Raster_Tab2(RasterHists *r);
+void Default_Setup_Raw_Raster_Tab3(RasterHists *r);
+void Default_Setup_1_Channel_Scope(RasterHists *r);
+void Default_Setup_2_Channel_Scope(RasterHists *r);
+void Default_Setup_3_Channel_Scope(RasterHists *r);
+void Default_Setup_Helicity(RasterHists *r);
+void Default_Initialize_Histograms(RasterHists *r);
+
+#endif //RASTERMON_RASTERMONDEFAULTCONFIG_H
